rename all 'members' with 'keys'
